+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const VERTICAL_EDITOR_TEXT = 'Open Editor (Next to the Active View)';
+export const HORIZONTAL_EDITOR_TEXT = 'Open Editor (Below the Active View)';
+export const POPOVER_EDITOR_TEXT = 'Open Editor (with the hover editor pluging)';
